add sqrt as function, still need to add simplify
