+++ src/main/java/io/github/atholton/sigmath/Topics/Topic.java
@@ -1,7 +1,7 @@
 package io.github.atholton.sigmath.Topics;
 import java.util.ArrayList;
 
-import io.github.atholton.sigmath.EquationTree.ASTNode;
+import io.github.atholton.sigmath.equationtree.ASTNode;
 
 /**
  * Represents an individual topic from the curriculum.
