+++ src/main/java/io/github/atholton/sigmath/equationtree/Operator.java
@@ -1,4 +1,4 @@
-package io.github.atholton.sigmath.EquationTree;
+package io.github.atholton.sigmath.equationtree;
 
 public interface Operator {
     /***
@@ -15,9 +15,7 @@ public interface Operator {
      * @param o The operator to compare against.
      *
      * @return -1 if this operator is of lower precedence, 1 if it's of greater
-     * precedence, and 0 if they're of equal precedence. If two operators are of
-     * equal precedence, right associativity and parenthetical groupings must be
-     * used to determine precedence.
+     * precedence, and 0 if they're of equal precedence.
      */
     int comparePrecedence(Operator o);
